cancel handling for row in edit mode

+++ src/SvelteGenericCrudTable.svelte
@@ -1,10 +1,18 @@
 <script>
     import {createEventDispatcher} from 'svelte';
+
     const dispatch = createEventDispatcher();
+    const EDIT = 'EDIT';
+    const DELETE = 'DELETE';
+    const DETAILS = 'DETAILS';
+    const NO_ROW_IN_EDIT_MODE = -1;
+
+    let row_in_edit_mode = NO_ROW_IN_EDIT_MODE;
 
     export let name = '';
     export let table = [];
     export let editable = [];
+    export let options = [EDIT, DELETE]
 
     function getKey(elem) {
         return elem.toString().split(',')[0];
@@ -19,7 +27,60 @@
         return elem.toString().split(',')[1];
     }
 
+    function resetEditmode(id) {
+        editable.forEach((toEdit) => {
+            document.getElementById(name + toEdit + id).setAttribute("disabled", "true");
+        })
+        document.getElementById(name + 'options-default' + id).classList.remove('hidden');
+        document.getElementById(name + 'options-default' + id).classList.add('shown');
+        document.getElementById(name + 'options-edit' + id).classList.remove('shown');
+        document.getElementById(name + 'options-edit' + id).classList.add('hidden');
+    }
+
+    function setEditmode(id) {
+        editable.forEach((toEdit) => {
+            document.getElementById(name + toEdit + id).removeAttribute("disabled");
+        })
+        document.getElementById(name + 'options-default' + id).classList.add('hidden');
+        document.getElementById(name + 'options-default' + id).classList.remove('shown');
+        document.getElementById(name + 'options-edit' + id).classList.remove('hidden');
+        document.getElementById(name + 'options-edit' + id).classList.add('shown');
+    }
+
+    function gatherUpdates(id) {
+        const body = {};
+        Object.entries(table[0]).forEach((elem) => {
+            body[getKey(elem)] = document.getElementById(name + getKey(elem) + id).value;
+        })
+        return body;
+    }
+
+    function resetRawInEditMode(id) {
+        if ((row_in_edit_mode !== id) && (row_in_edit_mode !== NO_ROW_IN_EDIT_MODE)) {
+            handleCancel(row_in_edit_mode);
+        }
+    }
+
+    function handleEdit(id) {
+        resetRawInEditMode(id);
+        row_in_edit_mode = id;
+        for (let i = 0; i < table.length; i++) {
+            resetEditmode(i);
+        }
+        setEditmode(id);
+    }
+
+    function handleUpdate(id) {
+        resetRawInEditMode(id);
+        const body = gatherUpdates(id);
+        dispatch('update', {
+            id: id,
+            body: body
+        });
+    }
+
     function handleDelete(id) {
+        resetRawInEditMode(id);
         const body = gatherUpdates(id);
         dispatch('delete', {
             id: id,
@@ -31,44 +92,18 @@
         dispatch('create', {});
     }
 
-    function resetEditmode(id) {
-        editable.forEach((toEdit) => {
-            document.getElementById(name + toEdit + id).setAttribute("disabled", "true");
-        })
-        document.getElementById(name + 'options' + id).classList.remove('hidden');
-        document.getElementById(name + 'options' + id).classList.add('shown');
-        document.getElementById(name + 'update' + id).classList.remove('shown');
-        document.getElementById(name + 'update' + id).classList.add('hidden');
+    function handleCancel(id) {
+        Object.entries(table[id]).forEach((elem) => {
+            document.getElementById(name + getKey(elem) + id).value = document.getElementById(name + getKey(elem) + id + 'copy').innerText;
+        });
+        resetEditmode(id);
+        row_in_edit_mode = NO_ROW_IN_EDIT_MODE;
     }
 
-    function setEditmode(id) {
-        editable.forEach((toEdit) => {
-            document.getElementById(name + toEdit + id).removeAttribute("disabled");
-        })
-        document.getElementById(name + 'options' + id).classList.add('hidden');
-        document.getElementById(name + 'options' + id).classList.remove('shown');
-        document.getElementById(name + 'update' + id).classList.remove('hidden');
-        document.getElementById(name + 'update' + id).classList.add('shown');
-    }
-
-    function handleEdit(id) {
-        for (let i = 0; i < table.length; i++) {
-            resetEditmode(i);
-        }
-        setEditmode(id);
-    }
-
-    function gatherUpdates(id) {
-        const body = {};
-        Object.entries(table[0]).forEach((elem) => {
-            body[getKey(elem)] = document.getElementById(name + getKey(elem) + id).value;
-        })
-        return body;
-    }
-
-    function handleUpdate(id) {
+    function handleDetails(id) {
+        resetRawInEditMode(id);
         const body = gatherUpdates(id);
-        dispatch('update', {
+        dispatch('details', {
             id: id,
             body: body
         });
@@ -91,16 +126,27 @@
                     <tr class="row">
                         {#each Object.entries(tableRow) as elem}
                             <td>
-                                <input id="{name}{getKey(elem)}{i}" value={getValue(elem)} disabled="true"/>
+                                <textarea id="{name}{getKey(elem)}{i}" value={getValue(elem)} disabled="true"/>
+                                <div class="hidden" id="{name}{getKey(elem)}{i}copy">{getValue(elem)}</div>
                             </td>
                         {/each}
                         <td>
-                            <div id="{name}options{i}" class="shown">
-                                <span class="options" on:click={() => handleDelete(i)}>delete</span>
-                                <span class="options" on:click={() => handleEdit(i)}>edit</span>
+                            <div id="{name}options-default{i}" class="shown">
+                                {#if options.includes(DELETE)}
+                                    <span class="options" on:click={() => handleDelete(i)}>delete</span>
+                                {/if}
+                                {#if options.includes(EDIT)}
+                                    <span class="options" on:click={() => handleEdit(i)}>edit</span>
+                                {/if}
+                                {#if options.includes(DETAILS)}
+                                    <span class="options" on:click={() => handleDetails(i)}>details</span>
+                                {/if}
                             </div>
-                            <div id="{name}update{i}" class="hidden">
-                                <span class="options" on:click={() => handleUpdate(i)}>update</span>
+                            <div id="{name}options-edit{i}" class="hidden">
+                                {#if options.includes(EDIT)}
+                                    <span class="options" on:click={() => handleUpdate(i)}>update</span>
+                                    <span class="options" on:click={() => handleCancel(i)}>cancel</span>
+                                {/if}
                             </div>
                         </td>
                     </tr>
@@ -109,7 +155,7 @@
         {:else}
             {JSON.stringify(table)}
         {/if}
-        <button class="button" on:click={handleCreate}>add {name}</button>
+        <button class="button" on:click={handleCreate}>add</button>
     {/if}
 </main>
 
@@ -126,6 +172,7 @@
         color: #5f5f5f;
         font-size: 0.85em;
         font-weight: 300;
+        padding-left: 0.2em;
     }
 
     td {
@@ -133,6 +180,7 @@
         border: none;
         font-size: 0.85em;
         font-weight: 200;
+        padding-left: 0.2em;
     }
 
     .button {
@@ -174,20 +222,27 @@
         color: #000000;
     }
 
-    input {
+    textarea {
         position: relative;
-        top: 0.3em;
+        resize: none;
+        overflow: auto;
+        top: 0.4em;
         width: 100%;
+        height: 1.3em;
+        padding: 1px;
         background-color: #ffffff;
         border: none;
         font-size: 0.85em;
         font-weight: 300;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+        overflow: hidden;
         -webkit-transition: box-shadow 0.3s;
         transition: box-shadow 0.3s;
-        box-shadow: -9px 9px 0px -8px #aaaaaa, 9px 9px 0px -8px #aaaaaa;
+        box-shadow: -6px 6px 0px -5px #aaaaaa, 6px 6px 0px -5px #aaaaaa;
     }
 
-    input:disabled {
+    textarea:disabled {
         color: #5f5f5f;
         background-color: inherit;
         font-size: 0.85em;
@@ -196,10 +251,11 @@
     }
 
     input:focus,
+    textarea:focus,
     select:focus {
         outline: none;
         font-weight: 300;
-        box-shadow: -9px 9px 0px -8px #5f5f5f, 9px 9px 0px -8px#5f5f5f;
+        box-shadow: -6px 6px 0px -5px #5f5f5f, 6px 6px 0px -5px #5f5f5f;
     }
 
     button:focus {
